adjusted chart value formatter
